+++ storage/backends/plakard/client.go
@@ -351,7 +351,7 @@ func (repository *Repository) GetStates() ([][32]byte, error) {
 	return result.Payload.(network.ResGetStates).Checksums, nil
 }
 
-func (repository *Repository) PutState(checksum [32]byte, rd io.Reader, size int64) error {
+func (repository *Repository) PutState(checksum [32]byte, rd io.Reader, size uint64) error {
 	data, err := io.ReadAll(rd)
 	if err != nil {
 		return err
@@ -370,7 +370,7 @@ func (repository *Repository) PutState(checksum [32]byte, rd io.Reader, size int
 	return nil
 }
 
-func (repository *Repository) GetState(checksum [32]byte) (io.Reader, int64, error) {
+func (repository *Repository) GetState(checksum [32]byte) (io.Reader, uint64, error) {
 	result, err := repository.sendRequest("ReqGetState", network.ReqGetState{
 		Checksum: checksum,
 	})
@@ -382,7 +382,7 @@ func (repository *Repository) GetState(checksum [32]byte) (io.Reader, int64, err
 		return nil, 0, fmt.Errorf("%s", result.Payload.(network.ResGetState).Err)
 	}
 
-	size := int64(len(result.Payload.(network.ResGetState).Data))
+	size := uint64(len(result.Payload.(network.ResGetState).Data))
 
 	return bytes.NewBuffer(result.Payload.(network.ResGetState).Data), size, nil
 }
@@ -414,7 +414,7 @@ func (repository *Repository) GetPackfiles() ([][32]byte, error) {
 
 }
 
-func (repository *Repository) PutPackfile(checksum [32]byte, rd io.Reader, size int64) error {
+func (repository *Repository) PutPackfile(checksum [32]byte, rd io.Reader, size uint64) error {
 	data, err := io.ReadAll(rd)
 	if err != nil {
 		return err
@@ -432,7 +432,7 @@ func (repository *Repository) PutPackfile(checksum [32]byte, rd io.Reader, size 
 	return nil
 }
 
-func (repository *Repository) GetPackfile(checksum [32]byte) (io.Reader, int64, error) {
+func (repository *Repository) GetPackfile(checksum [32]byte) (io.Reader, uint64, error) {
 	result, err := repository.sendRequest("ReqGetPackfile", network.ReqGetPackfile{
 		Checksum: checksum,
 	})
@@ -444,12 +444,12 @@ func (repository *Repository) GetPackfile(checksum [32]byte) (io.Reader, int64, 
 	}
 
 	data := result.Payload.(network.ResGetPackfile).Data
-	datalen := int64(len(data))
+	datalen := uint64(len(data))
 
 	return bytes.NewBuffer(data), datalen, nil
 }
 
-func (repository *Repository) GetPackfileBlob(checksum [32]byte, offset uint32, length uint32) (io.Reader, int64, error) {
+func (repository *Repository) GetPackfileBlob(checksum [32]byte, offset uint32, length uint32) (io.Reader, uint32, error) {
 	result, err := repository.sendRequest("ReqGetPackfileBlob", network.ReqGetPackfileBlob{
 		Checksum: checksum,
 		Offset:   offset,
@@ -464,7 +464,7 @@ func (repository *Repository) GetPackfileBlob(checksum [32]byte, offset uint32, 
 	}
 
 	data := result.Payload.(network.ResGetPackfileBlob).Data
-	datalen := int64(len(data))
+	datalen := uint32(len(data))
 
 	return bytes.NewBuffer(data), datalen, nil
 }
